Integrated AE2WTL More hei support Update API.md

+++ src/main/java/appeng/api/stacks/FuzzySearch.java
@@ -9,6 +9,7 @@ import java.util.Comparator;
 import java.util.SortedMap;
 
 final class FuzzySearch {
+
     @VisibleForTesting
     static final KeyComparator COMPARATOR = new KeyComparator();
     /**
@@ -143,7 +144,7 @@ final class FuzzySearch {
             // damage values to be predictable, while still having to satisfy the
             // complete order requirements of the sorted map
             // (We hope there won't be hash collisions... the probability is very low anyway)
-            return Long.compare(stackA.hashCode(), stackB.hashCode());
+            return Integer.compare(stackA.hashCode(), stackB.hashCode());
         }
     }
 }
